requirements.txt
streamlit>=1.28.0
python-dotenv>=0.19.0
requests>=2.31.0
tenacity>=8.2.3
pydantic>=2.5.0
Pillow>=10.1.0